Backend/Connection/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
qdrant-client>=1.9.0
numpy>=1.24.0
pandas>=2.0.0
openai==1.3.5
python-dotenv>=1.0.0
google-generativeai==0.4.0
cohere==5.20.0